Update html and switch to nginx

+++ start.sh
@@ -1,12 +1,20 @@
 #!/bin/sh
 
 # setup list of monitored ups hosts
-cp /etc/nut/hosts.conf.original /etc/nut/hosts.conf
+if [ -f /etc/nut/hosts.conf.original ]; then
+	cp /etc/nut/hosts.conf.original /etc/nut/hosts.conf
+	
+	IFS=';'
+	for host in $NUT_HOSTS
+	do
+	    echo "${host}" >> /etc/nut/hosts.conf
+	done
+fi
 
-IFS=';'
-for host in $NUT_HOSTS
-do
-    echo "${host}" >> /etc/nut/hosts.conf
-done
+# run the fcgiwrap daemon
+printf "Starting up the fcgiwrap daemon ...\n"
+service fcgiwrap start || { printf "ERROR on daemon startup.\n"; exit; }
 
-/usr/sbin/lighttpd -D -f /etc/lighttpd/lighttpd.conf
+# run nginx
+printf "Starting up the web server ...\n"
+exec nginx -g 'daemon off;'
